Move exception to separate file, import
